fix(tests): update sentinel imports from _shared to _sentinels after remote refactor
